Добавил соединение с беком на экране прохождения теста

+++ app/src/main/java/com/farma/parkinsoftapp/presentation/navigation/Destinations.kt
@@ -21,7 +21,7 @@ object PatientAllTestsRoute
 
 @Serializable
 data class PatientTestRoute(
-    val testId: Int,
+    val testId: Long,
     val testType: TestType
 )
